% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciecam16.R
\name{CIECAM16}
\alias{CIECAM16}
\title{CIECAM16 Color Appearance Model}
\description{
Implementation of CIECAM16 color appearance model for color space conversions and
color difference calculations. Supports:
- Transformations between CIE XYZ and CIECAM16
- Color difference calculations in CAM16-UCS space
- Multiple viewing conditions
}
\references{
CIE 248:2022. The CIE 2016 Colour Appearance Model for Colour Management Systems:
CIECAM16. Vienna: CIE Central Bureau.
}
